// src/types/planets.ts
type Planet = {
  id: number | undefined | null;
  name: string;
  diameter: string;
  rotation_period: string;
  orbital_period: string;
  gravity: string;
  population: string;
  climate: string;
  terrain: string;
  surface_water: string;
  created: string | null | undefined;
  edited: string | null | undefined;
};

export default Planet;